feat(runtime): delete the old gateway-catalog resolver chain (A9)
Deletes catalog::gateway_resolver.rs (the sqlite-backed resolver chain:
GatewayModelResolver, GatewayProbeStore, gateway_model_probe table) now
that the model-snapshot document + GatewayModelPlanner fully replace it.
Relocates the surviving pure functions (native_default_model,
provider_for_model, normalize_model_family) into a new projection.rs
module alongside enrich_model/resolve_catalog_match, which move there
from agent_gateway_catalog.rs.

Ruling (spec §3): keep the legacy gateway-catalog route URLs
(GET .../catalog/gateway-models, POST .../catalog/refresh-gateway)
stable for the desktop All-Models tab, but swap their backend to read
from the model-snapshot service instead of the deleted resolver. Both
routes are now thin readers/pokes over ModelSnapshotService.

Adds migration 0066 to drop the now-orphaned gateway_model_probe table
and its index (0051 stays, append-only history). Regenerates the SDK
(openapi.json/openapi.ts) and the checked-in schema snapshot to match.
