+++ scripts/max_lines_allowlist.txt
@@ -15,12 +15,11 @@ anyharness/crates/anyharness-lib/src/domains/materialization/service.rs 625 loca
 anyharness/crates/anyharness-lib/src/domains/workflows/service.rs 664 workflow durable rules (+run-control cancel-intent use case and helpers); split into service/ on next growth per guides/domains.md
 anyharness/crates/anyharness-lib/src/domains/workflows/service_tests.rs 1320 workflow-runs merge-gated domain suite (+2b partial-uniqueness contract on nonterminal session binding)
 anyharness/crates/anyharness-lib/src/api/workflow_runs_tests.rs 2790 workflow-runs HTTP + runtime/extension crossing suite (run-control PROOF-01 battery + 2b admission races: reservation/create/bind window, terminal release, PR1227-WORKSPACE-FENCE-01 purge/retire destruction-vs-late-bind proofs, and FENCE-02 bind->terminalize->destruction race proofs over the real executor)
-anyharness/crates/anyharness-lib/src/api/http/agent_gateway_catalog.rs 637 gateway model enrichment (own vs foreign match) + catalog version endpoint (+B5 doc-comment fix: GatewayModelsResponse.models / RefreshGatewayResponse.models no longer describe client-side provider filtering)
 anyharness/crates/anyharness-lib/src/api/http/workspaces_lifecycle.rs 645 workspace retire lifecycle handlers (+PR1227-WORKSPACE-FENCE-01/02 under-exclusive-lease workflow-control re-check with admitted-set fail-closed membership + test-only proof seam module)
 anyharness/crates/anyharness-lib/src/api/http/workspaces_purge.rs 665 workspace purge handlers (+2b fail-closed all-session admission, PR1227-WORKSPACE-FENCE-01/02 admitted-set carry + under-exclusive-lease re-check, test-only purge proof seam module)
 anyharness/crates/anyharness-lib/src/domains/workspaces/retention.rs 823 worktree retention sweep (+PR1227-RETENTION-FENCE-01 permit/lease fencing + test proof seam, +Workflow creator-context exclusion until an explicit cleanup policy exists)
 anyharness/crates/anyharness-lib/src/api/session_admission_tests.rs 674 2b admission conflict-matrix + fail-closed purge/retire/mobility-export/destroy-source proofs + LOCK-01 permit-before-lease deadlock/source-order guards (+idempotent create)
-anyharness/crates/anyharness-lib/src/domains/agents/catalog/gateway_resolver.rs 770 family normalizer (bedrock/date/version forms) with tests (+native-default-model resolution and its tests, incl. the opencode any-non-gateway-fallback pin); the launch-probe trigger and its credential/record helpers are gone with the poke wiring, and the client-side provider filter is gone with B5; the normalizer relocates with the old-chain deletion
+anyharness/crates/anyharness-lib/src/domains/agents/catalog/projection.rs 998 family normalizer (bedrock/date/version forms) with tests (+native-default-model resolution and its tests, incl. the opencode any-non-gateway-fallback pin), plus the enrichment join (own vs foreign catalog match) and its tests relocated from agent_gateway_catalog.rs under A9's route-backend swap (ruling §3): gateway_resolver.rs is deleted (old resolver chain), and the route file itself dropped to 312 lines and no longer needs an allowlist entry
 anyharness/crates/anyharness-lib/src/domains/agents/catalog/validation.rs 609 existing catalog validator router; unattended-mode invariants live in a focused submodule
 anyharness/crates/anyharness-lib/src/domains/agents/auth/context_tests.rs 645 classifier tests incl. gateway route activation + bedrock flag-only detection
 anyharness/crates/anyharness-lib/src/domains/cowork/runtime.rs 1893 existing AnyHarness oversized file (catalog-owned unattended default replaces local fallback helper)
@@ -63,7 +62,7 @@ apps/packages/product-client/src/lib/domain/workspaces/sidebar/sidebar.test.ts 6
 apps/packages/product-client/src/lib/domain/workspaces/sidebar/sidebar-indicators.test.ts 648 existing desktop oversized test file
 apps/packages/product-client/src/lib/infra/measurement/measurement-port.ts 734 product-owned measurement port barrel (R1 ruling) mirroring the retained measurement surface through a swappable no-op sink; split deferred
 anyharness/crates/anyharness-lib/src/api/router.rs 640 AnyHarness router (+goals/loops/activity/feed routes, +gateway catalog version route, +workflow-runs/cancel + local materialization operations + workflow-run-workspaces routes + worktree restore route)
-anyharness/crates/anyharness-lib/src/app/mod.rs 708 AnyHarness app wiring (+gateway model planner and probe-engine poke wiring) (+session mutation admission purge/retention wiring, +goal/loop/activity/feed services, +two-phase workflow-run and workflow-workspace wiring, +local materialization service wiring + worktree restore runtime wiring)
+anyharness/crates/anyharness-lib/src/app/mod.rs 697 AnyHarness app wiring (+gateway model planner wiring; the gateway model resolver + probe-engine store wiring is gone with the old-chain deletion, A9) (+session mutation admission purge/retention wiring, +goal/loop/activity/feed services, +two-phase workflow-run and workflow-workspace wiring, +local materialization service wiring + worktree restore runtime wiring)
 anyharness/crates/anyharness-lib/src/domains/loops/runtime.rs 867 loops runtime — native write path + emulated Codex scheduler (+2b derived-safe classification comment at the fire seam)
 anyharness/crates/anyharness-lib/src/live/sessions/actor/startup.rs 642 session startup (+goal/loop capability parsing, roster reconcile-on-attach)
 server/tests/integration/test_organizations_api.py 624 added admin-only coverage for GET .../invitations
